auth-service/requirements.txt
pytest==7.4.3
requests==2.31.0
pytest-html==4.1.1
pytest-json-report==1.5.0
pytest-xdist==3.5.0
python-dotenv==1.0.1
pytest-cov==4.1.0
